Change float to uint32 type values that represent width and height => no sense for me floating point in 2d sizes.

+++ vongine/src/vongine/physics/VGCircleSimple2DCollision.cpp
@@ -6,7 +6,7 @@
 
 NS_VG_BEGIN
 
-void CircleSimple2DCollision::UpdateShape(const float x, const float y, const float radius) 
+void CircleSimple2DCollision::UpdateShape(const float x, const float y, const uint32 radius)
 { 
 	_x = x; 
 	_y = y;
